Changed `SourceRectangle` to use `i32`
Closes https://github.com/PistonDevelopers/rust-graphics/issues/491

+++ src/color_context.rs
@@ -184,7 +184,7 @@ impl<'a, 'b, I: Image> AddImage<'a, 'b, ImageColorContext<'a, 'b, I>, I> for Col
             view: Borrowed(self.view.get()),
             transform: Borrowed(self.transform.get()),
             image: Value(image),
-            source_rect: Value([0, 0, w, h]),
+            source_rect: Value([0, 0, w as i32, h as i32]),
             color: Borrowed(self.color.get()),
         }
     }
